requirements.txt
streamlit
llama-parse
llama-index
llama-index-core
google-generativeai
llama-index-llms-gemini
llama-index-embeddings-gemini
python-dotenv
pinecone-client
llama-index-vector-stores-pinecone
pydantic
streamlit
llama-index-llms-azure-openai
llama-index-embeddings-openai